Change by comments, add deletation of silent audio files
