+++ server/services/process-chunk.service.ts
@@ -1,8 +1,10 @@
 import {Class, IChunkScheme, Status} from '../models/chunk.model';
+import { hasAudibleSound } from './audibleSoundDetector';
 import {ChunkService} from './chunk.service';
 import {analyzeAudio} from './speechmatics.service';
-import {AnalysisResult, analyzeToneAndWords, EmptyResponse} from "./transcribe-analyzer.service";
+import {AnalysisResult, analyzeToneAndWords} from "./transcribe-analyzer.service";
 import { RetrieveTranscriptResponseAlternative } from './transcribe-analyzer.service';
+import { promises as fs } from 'fs';
 
 export function getRandomStatus(): Status {
     const statuses = [Status.NotStarted, Status.InProgress, Status.Completed];
@@ -24,18 +26,28 @@ export async function processChunk(chunk: IChunkScheme) {
         // Update chunk status to in-progress
         await ChunkService.updateChunk(chunk.id, { status: Status.InProgress });
 
-        // Send audio to Speechmatics
-        const result: RetrieveTranscriptResponseAlternative | EmptyResponse = await analyzeAudio(chunk.audioFilePath);
-
-        // Handle empty chunks (silent or under threshold)
-        if ('emptyChunk' in result && result.emptyChunk) {
+        const hasSound = await hasAudibleSound(chunk.audioFilePath);
+        if (!hasSound) {
+            console.log('Audio is below threshold. Skipping transcription.');
+     
             await ChunkService.updateChunk(chunk.id, {
                 status: Status.Completed,
                 chunkClass: Class.Natural,
                 summary: 'No meaningful audio detected',
             });
+
+            try {
+                await fs.unlink(chunk.audioFilePath);
+                console.log('File deleted:', chunk.audioFilePath);
+            } catch (err) {
+                console.error('Failed to delete file:', err);
+            }
+
             return;
         }
+
+        // Send audio to Speechmatics
+        const result: RetrieveTranscriptResponseAlternative = await analyzeAudio(chunk.audioFilePath);
 
         // Analyze result for tone and bad words
         const analysisResult: AnalysisResult = analyzeToneAndWords(result as RetrieveTranscriptResponseAlternative);
