+++ server/services/transcribe-analyzer.service.ts
@@ -23,10 +23,6 @@ export type RetrieveTranscriptResponseAlternative = {
     };
 }
 
-export type EmptyResponse = {
-    emptyChunk: boolean
-}
-
 export function analyzeToneAndWords(response: RetrieveTranscriptResponseAlternative): AnalysisResult {
     const summary = response?.sentiment_analysis?.summary?.overall || response?.sentiment_analysis?.sentiment_analysis?.summary?.overall;
